requirements.txt
pydantic
PyYAML
jinja2
beautifulsoup4
pytest
